server.js refactor to accomodate firebase functions
