+++ .cache/02/9811ca274e2d41cc5e3c2129ab80f6.json
@@ -1,0 +1,1 @@
+{"id":"lw9M","dependencies":[{"name":"/Users/ryan.tercias/MyApps/territory-data-service/package.json","includedInParent":true,"mtime":1592367476738},{"name":"../server","loc":{"line":1,"column":21},"parent":"/Users/ryan.tercias/MyApps/territory-data-service/src/async/reports.js","resolved":"/Users/ryan.tercias/MyApps/territory-data-service/src/server.js"}],"generated":{"js":"\"use strict\";Object.defineProperty(exports,\"__esModule\",{value:!0}),exports.default=void 0;var e=require(\"../server\"),r=require(\"lodash\");class t{async getAssignmentRecords(t){return(0,r.toArray)(await e.conn.query(`SELECT * FROM territorycheckouts_pivot where congregationid=${t} order by territory_id, timestamp`))}}var o=new t;exports.default=o;"},"sourceMaps":{"js":{"mappings":[{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":0}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":13}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":20}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":35}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":43}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":56}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":57}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":64}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":68}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":76}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":89}},{"source":"async/reports.js","original":{"line":1,"column":0},"generated":{"line":1,"column":91}},{"source":"async/reports.js","original":{"line":1,"column":0},"generated":{"line":1,"column":95}},{"source":"async/reports.js","original":{"line":1,"column":0},"generated":{"line":1,"column":97}},{"source":"async/reports.js","original":{"line":1,"column":0},"generated":{"line":1,"column":105}},{"source":"async/reports.js","original":{"line":2,"column":0},"generated":{"line":1,"column":118}},{"source":"async/reports.js","original":{"line":2,"column":0},"generated":{"line":1,"column":120}},{"source":"async/reports.js","original":{"line":2,"column":0},"generated":{"line":1,"column":128}},{"source":"async/reports.js","original":{"line":4,"column":0},"generated":{"line":1,"column":138}},{"source":"async/reports.js","name":"ReportsAsync","original":{"line":4,"column":6},"generated":{"line":1,"column":144}},{"source":"async/reports.js","name":"getAssignmentRecords","original":{"line":5,"column":8},"generated":{"line":1,"column":146}},{"source":"async/reports.js","name":"congId","original":{"line":5,"column":29},"generated":{"line":1,"column":173}},{"source":"async/reports.js","original":{"line":6,"column":11},"generated":{"line":1,"column":176}},{"source":"async/reports.js","original":{"line":6,"column":11},"generated":{"line":1,"column":183}},{"source":"async/reports.js","original":{"line":6,"column":19},"generated":{"line":1,"column":185}},{"source":"async/reports.js","original":{"line":6,"column":19},"generated":{"line":1,"column":187}},{"source":"async/reports.js","name":"conn","original":{"line":6,"column":25},"generated":{"line":1,"column":202}},{"source":"async/reports.js","name":"query","original":{"line":6,"column":30},"generated":{"line":1,"column":204}},{"source":"async/reports.js","name":"query","original":{"line":6,"column":30},"generated":{"line":1,"column":209}},{"source":"async/reports.js","name":"congId","original":{"line":6,"column":99},"generated":{"line":1,"column":278}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":318}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":322}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":324}},{"source":"async/reports.js","name":"ReportsAsync","original":{"line":10,"column":19},"generated":{"line":1,"column":328}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":330}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":338}},{"source":"async/reports.js","original":{"line":10,"column":15},"generated":{"line":1,"column":346}}],"sources":{"async/reports.js":"import { conn } from '../server';\nimport { toArray } from 'lodash';\n\nclass ReportsAsync {\n  async getAssignmentRecords(congId) {\n    return toArray(await conn.query(`SELECT * FROM territorycheckouts_pivot where congregationid=${congId} order by territory_id, timestamp`));\n  }\n}\n\nexport default new ReportsAsync();\n"},"lineCount":null}},"error":null,"hash":"b8687e0a46e61d4010b595af4c1a95da","cacheData":{"env":{}}}
