+++ .cache/13/3188a09006adac411eff7a9f8fb2a2.json
@@ -1,0 +1,1 @@
+{"id":"Oc2D","dependencies":[{"name":"/Users/ryan.tercias/MyApps/territory-data-service/package.json","includedInParent":true,"mtime":1592367476738},{"name":"../server","loc":{"line":2,"column":21},"parent":"/Users/ryan.tercias/MyApps/territory-data-service/src/async/congregations.js","resolved":"/Users/ryan.tercias/MyApps/territory-data-service/src/server.js"}],"generated":{"js":"\"use strict\";Object.defineProperty(exports,\"__esModule\",{value:!0}),exports.default=void 0;var e=require(\"lodash\"),r=require(\"../server\");class n{async getCongregationById(e){return(await r.conn.query(`SELECT * FROM congregations WHERE id=${e}`))[0]}async getAllCongregations(){return(0,e.toArray)(await r.conn.query(\"SELECT * FROM congregations\"))}async searchCongregations(n){return(0,e.toArray)(await r.conn.query(`SELECT * FROM congregations WHERE name LIKE '%${n}%' OR description LIKE '%${n}%'`))}}var a=new n;exports.default=a;"},"sourceMaps":{"js":{"mappings":[{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":0}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":13}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":20}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":35}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":43}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":56}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":57}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":64}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":68}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":76}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":89}},{"source":"async/congregations.js","original":{"line":1,"column":0},"generated":{"line":1,"column":91}},{"source":"async/congregations.js","original":{"line":1,"column":0},"generated":{"line":1,"column":95}},{"source":"async/congregations.js","original":{"line":1,"column":0},"generated":{"line":1,"column":97}},{"source":"async/congregations.js","original":{"line":1,"column":0},"generated":{"line":1,"column":105}},{"source":"async/congregations.js","original":{"line":2,"column":0},"generated":{"line":1,"column":115}},{"source":"async/congregations.js","original":{"line":2,"column":0},"generated":{"line":1,"column":117}},{"source":"async/congregations.js","original":{"line":2,"column":0},"generated":{"line":1,"column":125}},{"source":"async/congregations.js","original":{"line":4,"column":0},"generated":{"line":1,"column":138}},{"source":"async/congregations.js","name":"CongregationAsync","original":{"line":4,"column":6},"generated":{"line":1,"column":144}},{"source":"async/congregations.js","name":"getCongregationById","original":{"line":5,"column":8},"generated":{"line":1,"column":146}},{"source":"async/congregations.js","name":"id","original":{"line":5,"column":29},"generated":{"line":1,"column":172}},{"source":"async/congregations.js","original":{"line":6,"column":11},"generated":{"line":1,"column":175}},{"source":"async/congregations.js","name":"conn","original":{"line":6,"column":18},"generated":{"line":1,"column":188}},{"source":"async/congregations.js","name":"query","original":{"line":6,"column":23},"generated":{"line":1,"column":190}},{"source":"async/congregations.js","name":"query","original":{"line":6,"column":23},"generated":{"line":1,"column":195}},{"source":"async/congregations.js","name":"id","original":{"line":6,"column":69},"generated":{"line":1,"column":241}},{"source":"async/congregations.js","original":{"line":6,"column":76},"generated":{"line":1,"column":247}},{"source":"async/congregations.js","name":"getAllCongregations","original":{"line":9,"column":8},"generated":{"line":1,"column":250}},{"source":"async/congregations.js","original":{"line":10,"column":11},"generated":{"line":1,"column":278}},{"source":"async/congregations.js","original":{"line":10,"column":11},"generated":{"line":1,"column":285}},{"source":"async/congregations.js","original":{"line":10,"column":19},"generated":{"line":1,"column":287}},{"source":"async/congregations.js","original":{"line":10,"column":19},"generated":{"line":1,"column":289}},{"source":"async/congregations.js","name":"conn","original":{"line":10,"column":25},"generated":{"line":1,"column":304}},{"source":"async/congregations.js","name":"query","original":{"line":10,"column":30},"generated":{"line":1,"column":306}},{"source":"async/congregations.js","name":"query","original":{"line":10,"column":30},"generated":{"line":1,"column":311}},{"source":"async/congregations.js","original":{"line":10,"column":37},"generated":{"line":1,"column":317}},{"source":"async/congregations.js","name":"searchCongregations","original":{"line":13,"column":8},"generated":{"line":1,"column":349}},{"source":"async/congregations.js","name":"keyword","original":{"line":13,"column":29},"generated":{"line":1,"column":375}},{"source":"async/congregations.js","original":{"line":14,"column":11},"generated":{"line":1,"column":378}},{"source":"async/congregations.js","original":{"line":14,"column":11},"generated":{"line":1,"column":385}},{"source":"async/congregations.js","original":{"line":14,"column":19},"generated":{"line":1,"column":387}},{"source":"async/congregations.js","original":{"line":14,"column":19},"generated":{"line":1,"column":389}},{"source":"async/congregations.js","name":"conn","original":{"line":14,"column":25},"generated":{"line":1,"column":404}},{"source":"async/congregations.js","name":"query","original":{"line":14,"column":30},"generated":{"line":1,"column":406}},{"source":"async/congregations.js","name":"query","original":{"line":14,"column":30},"generated":{"line":1,"column":411}},{"source":"async/congregations.js","name":"keyword","original":{"line":14,"column":85},"generated":{"line":1,"column":466}},{"source":"async/congregations.js","name":"keyword","original":{"line":14,"column":120},"generated":{"line":1,"column":495}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":504}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":508}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":510}},{"source":"async/congregations.js","name":"CongregationAsync","original":{"line":18,"column":19},"generated":{"line":1,"column":514}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":516}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":524}},{"source":"async/congregations.js","original":{"line":18,"column":15},"generated":{"line":1,"column":532}}],"sources":{"async/congregations.js":"import { toArray } from 'lodash';\nimport { conn } from '../server';\n\nclass CongregationAsync {\n  async getCongregationById (id) {\n    return (await conn.query(`SELECT * FROM congregations WHERE id=${id}`))[0];\n  }\n\n  async getAllCongregations () {\n    return toArray(await conn.query(`SELECT * FROM congregations`));\n  }\n\n  async searchCongregations (keyword) {\n    return toArray(await conn.query(`SELECT * FROM congregations WHERE name LIKE '%${keyword}%' OR description LIKE '%${keyword}%'`));\n  }\n}\n\nexport default new CongregationAsync();\n"},"lineCount":null}},"error":null,"hash":"dd042d06ab0df5e725ad5bea31841083","cacheData":{"env":{}}}
