+++ .cache/d9/3a65e5730dabcfc0317396bb6e0367.json
@@ -1,0 +1,1 @@
+{"id":"guaG","dependencies":[{"name":"/Users/ryan.tercias/MyApps/territory-data-service/package.json","includedInParent":true,"mtime":1592367476738},{"name":"./server","loc":{"line":2,"column":26},"parent":"/Users/ryan.tercias/MyApps/territory-data-service/src/functions.js","resolved":"/Users/ryan.tercias/MyApps/territory-data-service/src/server.js"}],"generated":{"js":"\"use strict\";Object.defineProperty(exports,\"__esModule\",{value:!0}),exports.api=void 0;var e=require(\"firebase-functions\"),r=require(\"./server\");const s=(0,r.gqlServer)(),t=e.https.onRequest(s);exports.api=t;"},"sourceMaps":{"js":{"mappings":[{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":0}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":13}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":20}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":35}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":43}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":56}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":57}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":64}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":68}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":76}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":85}},{"source":"functions.js","original":{"line":1,"column":0},"generated":{"line":1,"column":87}},{"source":"functions.js","original":{"line":1,"column":0},"generated":{"line":1,"column":91}},{"source":"functions.js","original":{"line":1,"column":0},"generated":{"line":1,"column":93}},{"source":"functions.js","original":{"line":1,"column":0},"generated":{"line":1,"column":101}},{"source":"functions.js","original":{"line":2,"column":0},"generated":{"line":1,"column":123}},{"source":"functions.js","original":{"line":2,"column":0},"generated":{"line":1,"column":125}},{"source":"functions.js","original":{"line":2,"column":0},"generated":{"line":1,"column":133}},{"source":"functions.js","original":{"line":4,"column":0},"generated":{"line":1,"column":145}},{"source":"functions.js","name":"server","original":{"line":4,"column":6},"generated":{"line":1,"column":151}},{"source":"functions.js","original":{"line":4,"column":15},"generated":{"line":1,"column":154}},{"source":"functions.js","original":{"line":4,"column":0},"generated":{"line":1,"column":156}},{"source":"functions.js","original":{"line":4,"column":0},"generated":{"line":1,"column":158}},{"source":"functions.js","name":"api","original":{"line":8,"column":6},"generated":{"line":1,"column":171}},{"source":"functions.js","name":"https","original":{"line":8,"column":12},"generated":{"line":1,"column":173}},{"source":"functions.js","name":"onRequest","original":{"line":8,"column":18},"generated":{"line":1,"column":175}},{"source":"functions.js","name":"onRequest","original":{"line":8,"column":18},"generated":{"line":1,"column":181}},{"source":"functions.js","name":"server","original":{"line":8,"column":28},"generated":{"line":1,"column":191}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":194}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":202}},{"source":"functions.js","original":{"line":8,"column":0},"generated":{"line":1,"column":206}}],"sources":{"functions.js":"import { https } from 'firebase-functions';\nimport { gqlServer } from './server';\n\nconst server = gqlServer();\n\n// Graphql api\n// https://us-east1-territory-data-service.cloudfunctions.net/api/\nconst api = https.onRequest(server);\n\nexport {api};\n"},"lineCount":null}},"error":null,"hash":"145c585521b486a2cc40e0e20b8c8183","cacheData":{"env":{}}}
