+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - 12.16.3
+  - 10
 cache:
   directories:
   - "$HOME/.npm"
@@ -19,4 +19,4 @@ env:
   CERTIFICATE_FILE:
     secure:
 script:
-  - npm run build
+  - yarn build
