+++ firebase.json
@@ -1,0 +1,1 @@
+{ "functions": { "source": "." } }
