+++ package.json
@@ -2,14 +2,21 @@
   "name": "territory-data-service",
   "version": "1.0.0",
   "description": "graphQL service",
-  "main": "index.js",
-  "engine": {
-    "node": ">=12.16"
+  "main": "dist/functions.js",
+  "engines": {
+    "node": "10"
   },
   "scripts": {
-    "build": "rm -rf dist/ && npx babel src --out-dir dist --ignore node_modules,.babel.config.js,package.json,npm-debug.log --copy-files",
-    "start": "nodemon --watch src/ --exec  \"npm run build && node dist/index.js\"",
-    "serve": "npm run build && node dist/index.js"
+    "lint": "xo",
+    "watch": "parcel src/functions.js --target node",
+    "build": "parcel build src/functions.js --target node --detailed-report",
+    "serve": "yarn firebase serve --only functions",
+    "emulator": "yarn firebase emulators:start --only functions",
+    "shell": "yarn firebase functions:shell",
+    "start": "yarn run shell",
+    "predeploy": "yarn build",
+    "deploy": "yarn firebase deploy --only functions",
+    "logs": "yarn firebase functions:log"
   },
   "repository": {
     "type": "git",
@@ -21,15 +28,18 @@
     "url": "https://github.com/rtercias/territory-data-service/issues"
   },
   "homepage": "https://github.com/rtercias/territory-data-service#readme",
+  "xo": {
+    "space": true
+  },
   "dependencies": {
-    "apollo-server-express": "^1.4.0",
-    "apollo-tracing": "^0.9.1",
-    "cluster": "^0.7.7",
-    "cors": "^2.8.5",
-    "date-fns": "^1.30.1",
+    "apollo-server-express": "^2.14.2",
+    "date-fns": "^2.14.0",
     "express": "^4.17.1",
-    "graphql": "^0.13.2",
-    "graphql-tools": "^3.1.1",
+    "firebase-admin": "^8.12.1",
+    "firebase-functions": "^3.7.0",
+    "firebase-tools": "8.4.2",
+    "graphql": "^14.3.0",
+    "graphql-tools": "^6.0.10",
     "mysql": "^2.18.1",
     "os": "^0.1.1",
     "promise-mysql": "^3.3.2",
@@ -43,7 +53,8 @@
     "babel-plugin-module-resolver": "^3.2.0",
     "chokidar": "^3.4.0",
     "eslint": "^4.19.1",
-    "fsevents": "^2.1.3",
-    "nodemon": "^1.19.4"
+    "firebase-functions-test": "^0.2.1",
+    "parcel-bundler": "^1.12.3",
+    "xo": "^0.24.0"
   }
 }
