+++ src/async/publishers.js
@@ -1,5 +1,5 @@
 import { toArray } from 'lodash';
-import { conn } from '../index';
+import { conn } from '../server';
 
 class PublisherAsync {
   async getPublisherById (id) {
