+++ src/server.js
@@ -1,0 +1,42 @@
+/**
+ * Copyright 2018 Territory Data Service
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), 
+ * to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+ * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES 
+ * OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE 
+ * OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import express from 'express';
+import mysql from 'mysql';
+import { ApolloServer } from 'apollo-server-express';
+import { promisify } from 'util';
+import { resolvers, typeDefs } from './schema/schema';
+
+export const conn = mysql.createPool({
+  connectionLimit: 10,
+  ssl: { rejectUnauthorized: false }, // TODO: add SSL certificate file here (see https://github.com/mysqljs/mysql#ssl-options)
+  host: process.env.TERRITORY_SERVER,
+  user: process.env.TERRITORY_USERID,
+  password: process.env.TERRITORY_PASSWORD,
+  database: 'territory'
+});
+
+export function gqlServer() {
+  const app = express();
+  const apolloServer = new ApolloServer({
+    typeDefs,
+    resolvers,
+    introspection: true,
+    playground: true,
+  });
+
+  conn.query = promisify(conn.query);
+  apolloServer.applyMiddleware({ app, path: '/', cors: true });
+  return app;
+}
